Modify the `reassign` command to rebalance partitions accross brokers whith the minimum number of partitions moves
BUG and TODO:
The migrate function can sometimes fail, because i try to move partitions from the biggest unbalanced node to the lowest unbalanced node, and sometimes there is not enough difference. Instead, i need to move remaining partitions that are already there on the other unbalanced nodes.
Nevertheless, if that happens, simply restart the script until the combination fits.